windows_dev_version/requirements.txt
altgraph==0.17.2
auto-py-to-exe==2.15.0
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
img2pdf==0.4.3
lxml==4.7.1
packaging==21.3
pdf2image==1.16.0
pefile==2021.9.3
pikepdf==4.3.1
Pillow==9.0.0
poppler-utils==0.1.0
pycparser==2.21
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
pyparsing==3.0.7
PyPDF4==1.27.0
PySide2==5.15.2.1
pywin32-ctypes==0.2.0
shiboken2==5.15.2.1
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
